requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
pandas>=2.0.0
tenacity>=8.0.0
